Use module for stepcounter notification via BLE
Avoid coupling bma and ble low-level modules.
